updated some things

+++ server/routes/booking.js
@@ -38,11 +38,19 @@ router.post("/", authMiddleware, async (req, res) => {
 router.delete("/:id", authMiddleware, async (req, res) => {
   try {
     const bookingId = req.params.id;
+    const userId = req.user._id;
+
+    const bookingToDelete = await Booking.findById(bookingId);
+    
+    if (bookingToDelete.user._id.toString() !== userId.toString()) {
+      return res.status(403).json({ message: "Unauthorized. You do not have permission to delete this booking." });
+    }
+
     const deletedBooking = await Booking.findByIdAndDelete(bookingId);
     if (!deletedBooking) {
       return res.status(404).json({ message: "Booking not found" });
     }
-    return res.status(200).json({ message: "Booking deleted successfully" });
+    return res.status(200).json({ message: "Booking canceled successfully" });
   } catch (error) {
     console.log(error);
     return res.status(500).json({ message: "Internal Server Error" });
